move export functionality into service and inject app date pipe
