+++ src/app/shared/services/grid-export.service.ts
@@ -1,0 +1,85 @@
+import { Injectable } from '@angular/core';
+import { NO_TEXT, YES_TEXT } from '../constants/constants';
+import { validateUrl } from '../helpers/link.helpers';
+import { GridColumn } from '../models/grid-column.model';
+import { AppDatePipe } from '../pipes/app-date.pipe';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class GridExportService {
+  constructor(private appDatePipe: AppDatePipe) {}
+
+  /**
+   * @param item An item representiung a row in the grid
+   * @param exportColumns The columns to be exported
+   * @returns The row with the column values transformed into something appropiate for the export
+   */
+  //eslint-disable-next-line @typescript-eslint/no-explicit-any
+  public transformRow(item: any, exportColumns: GridColumn[]): any {
+    const transformedItem = { ...item };
+    exportColumns.forEach((column) => {
+      const field = column.field;
+      if (field) {
+        switch (column.style) {
+          case 'chip':
+            if (typeof transformedItem[field] === 'boolean') {
+              const boolValue = transformedItem[field] ? 0 : 1;
+              transformedItem[field] = column.extraData[boolValue].name;
+            }
+            break;
+          case 'enum':
+            if (typeof transformedItem[field] === 'object') {
+              const enumValue = transformedItem[field];
+              transformedItem[field] = enumValue.name;
+            }
+            break;
+          case 'boolean':
+            {
+              transformedItem[field] = transformedItem[field] ? YES_TEXT : NO_TEXT;
+            }
+            break;
+          case 'uuid-to-name':
+            transformedItem[field] = transformedItem[`${column.dataField}`];
+            break;
+          case 'excel-only': {
+            if (transformedItem.RoleEmails) {
+              const roleEmailKeys: string[] = Object.keys(transformedItem.RoleEmails);
+              roleEmailKeys.forEach((key) => {
+                const prefixedKey = `Roles.${key}`;
+                if (prefixedKey === field) {
+                  transformedItem[`${column.title}`] = transformedItem.RoleEmails[key];
+                }
+              });
+            }
+            break;
+          }
+          case 'page-link-array':
+            {
+              const array = transformedItem[column.dataField as string];
+              const excelValue = array.map((item: { value: string }) => item.value).join(', ');
+              transformedItem[field] = excelValue;
+            }
+            break;
+          case 'usages':
+            {
+              const usages = transformedItem[column.field as string];
+              transformedItem[field] = usages.length;
+            }
+            break;
+          case 'title-link':
+            {
+              const url = transformedItem[column.idField ?? ''];
+              const exportValue = validateUrl(url) ? url : transformedItem[field];
+              transformedItem[field] = exportValue;
+            }
+            break;
+
+          default:
+            break;
+        }
+      }
+    });
+    return transformedItem;
+  }
+}
